removed css. fixed column scrolling. restructured app a bit after some advice from mxstbr about sharing styled-components.
