+++ npm-debug.log
@@ -1,0 +1,47 @@
+0 info it worked if it ends with ok
+1 verbose cli [ '/usr/local/Cellar/node/7.3.0/bin/node',
+1 verbose cli   '/usr/local/bin/npm',
+1 verbose cli   'start' ]
+2 info using npm@4.0.5
+3 info using node@v7.3.0
+4 verbose run-script [ 'prestart', 'start', 'poststart' ]
+5 info lifecycle Spectrum@0.1.0~prestart: Spectrum@0.1.0
+6 silly lifecycle Spectrum@0.1.0~prestart: no script for prestart, continuing
+7 info lifecycle Spectrum@0.1.0~start: Spectrum@0.1.0
+8 verbose lifecycle Spectrum@0.1.0~start: unsafe-perm in lifecycle true
+9 verbose lifecycle Spectrum@0.1.0~start: PATH: /usr/local/lib/node_modules/npm/bin/node-gyp-bin:/Users/bryn/Code/spectrum/node_modules/.bin:/bin:/sbin:/usr/bin:/usr/local/sbin:/usr/local/bin:./bin:/usr/local/bin:/usr/local/sbin:/Users/bryn/.dotfiles/bin:/Users/bryn/.rbenv/shims:/Users/bryn/.yarn/bin:/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin:/opt/X11/bin
+10 verbose lifecycle Spectrum@0.1.0~start: CWD: /Users/bryn/Code/spectrum
+11 silly lifecycle Spectrum@0.1.0~start: Args: [ '-c', 'react-scripts start' ]
+12 silly lifecycle Spectrum@0.1.0~start: Returned: code: 1  signal: null
+13 info lifecycle Spectrum@0.1.0~start: Failed to exec start script
+14 verbose stack Error: Spectrum@0.1.0 start: `react-scripts start`
+14 verbose stack Exit status 1
+14 verbose stack     at EventEmitter.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/lifecycle.js:279:16)
+14 verbose stack     at emitTwo (events.js:106:13)
+14 verbose stack     at EventEmitter.emit (events.js:191:7)
+14 verbose stack     at ChildProcess.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/spawn.js:40:14)
+14 verbose stack     at emitTwo (events.js:106:13)
+14 verbose stack     at ChildProcess.emit (events.js:191:7)
+14 verbose stack     at maybeClose (internal/child_process.js:885:16)
+14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:226:5)
+15 verbose pkgid Spectrum@0.1.0
+16 verbose cwd /Users/bryn/Code/spectrum
+17 error Darwin 16.1.0
+18 error argv "/usr/local/Cellar/node/7.3.0/bin/node" "/usr/local/bin/npm" "start"
+19 error node v7.3.0
+20 error npm  v4.0.5
+21 error code ELIFECYCLE
+22 error Spectrum@0.1.0 start: `react-scripts start`
+22 error Exit status 1
+23 error Failed at the Spectrum@0.1.0 start script 'react-scripts start'.
+23 error Make sure you have the latest version of node.js and npm installed.
+23 error If you do, this is most likely a problem with the Spectrum package,
+23 error not with npm itself.
+23 error Tell the author that this fails on your system:
+23 error     react-scripts start
+23 error You can get information on how to open an issue for this project with:
+23 error     npm bugs Spectrum
+23 error Or if that isn't available, you can get their info via:
+23 error     npm owner ls Spectrum
+23 error There is likely additional logging output above.
+24 verbose exit [ 1, true ]
