+++ public/index.html
@@ -4,9 +4,10 @@
     <meta charset="utf-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link rel="shortcut icon" href="%PUBLIC_URL%/img/favicon.ico">
+    <link rel="stylesheet" type="text/css" href="%PUBLIC_URL%/reset.css">
     <title>Spectrum</title>
   </head>
-  <body style="position: relative; overflow: hidden; height: 100vw; width: 100vw;">
-    <div style="height: 100vw; width: 100vw;" id="root"></div>
+  <body>
+    <div id="root"></div>
   </body>
 </html>
