+++ public/reset.css
@@ -1,0 +1,77 @@
+* {
+  border: 0;
+  box-sizing: border-box;
+  -webkit-font-smoothing: antialiased;
+  font-weight: normal;
+  margin: 0;
+  outline: 0;
+  padding: 0;
+  text-decoration: none;
+  text-rendering: optimizeLegibility;
+  // list-style: none;
+}
+
+html,
+body {
+  min-height: 100vh;
+  min-width: 100%;
+}
+
+a {
+  color: currentColor;
+  text-decoration: none;
+}
+
+textarea {
+  resize: none;
+}
+
+::selection {
+  background: #7D16FF;
+  color: #ffffff;
+}
+
+::-webkit-scrollbar { 
+    display: none; 
+}
+
+input,
+textarea {
+  box-shadow: inset 0 1px 1px rgba(0,0,0,0.1);
+}
+
+body {
+  position: relative;  
+  height: 100vh;
+  width: 100vw;
+  min-width: 768px;
+  -webkit-font-smoothing: antialiased;
+  font-size: 1rem;
+  line-height: 1rem;
+  background: #fff;
+  color: #555459;
+  padding: 0;
+  overflow-y: hidden;
+  overflow-x: hidden;
+  -webkit-user-select: none;
+  -moz-user-select: none;
+  -ms-user-select: none;
+  user-select: none;
+}
+
+#root {
+  display: flex;
+  display: -webkit-box;
+  display: -webkit-flex;
+  display: -moz-flex;
+  display: -ms-flexbox;
+  flex-direction: column;
+  -ms-flex-direction: column;
+  -moz-flex-direction: column;
+  -webkit-flex-direction: column;
+  height: 100%;
+  -webkit-user-select: none;
+  -moz-user-select: none;
+  -ms-user-select: none;
+  user-select: none;
+}
