+++ src/App/components/ListDetail/index.js
