+++ src/App/components/NavBar/svg.js
