+++ src/App/components/Post/index.js
@@ -1,5 +1,4 @@
 import React, { Component } from 'react';
-import posts from '../../data/posts.js';
 import { PostWrapper, PostBody, PostImg } from './style';
 import UserHeader from '../UserHeader'
 
@@ -9,7 +8,7 @@ class Post extends Component{
 	    	<PostWrapper>
 	    		<UserHeader />
 	    		<PostBody>
-	    			<p className="mt2 h5 color-primary">{posts[0].post.body.text}</p>
+	    			<p className="mt2 h5 color-primary"></p>
 	    			<PostImg src="/img/media.png" role="presentation"/>
 	    		</PostBody>
     		</PostWrapper>
