+++ src/App/components/PostList/style.js
@@ -8,16 +8,22 @@ export const Column = styled.div`
 	background-color: #F5F6F7;
 	height: 100%;
 	position: relative;
+	overflow-y: scroll;
 `;
 
 export const ScrollBody = styled.div`
-	overflow-y: scroll;
-	height: 100%;
 	position: relative;
+	z-index: 1;
+	padding-top: 56px;
 	`;
 
 export const ActionHeader = styled.div`
+	flex: 0 0 56px;
+	position: fixed;
 	background-color: #ffffff;
 	height: 56px;
+	width: 100%;
+	top: 0;
 	border-bottom: 1px solid rgba(0,0,0,0.1);
+	z-index: 2
 `;
