+++ src/App/components/UserHeader/svg.js
