+++ src/App/style.js
@@ -2,5 +2,7 @@ import styled from 'styled-components';
 
 export const Body = styled.div`
   display: flex;
+  overflow: hidden;
+  height: 100%;
   font-family: -apple-system, BlinkMacSystemFont, sans-serif;
 `;
