--- src/css/_background.scss
@@ -1,20 +1,0 @@
-.bg-wash {
-  background-color: $shade-gray;
-}
-
-.bg-default {
-  background-color: $white;
-}
-
-.bg-reverse {
-	background-color: $black;
-}
-
-.bg-primary {background-color: $spectre-gray}
-
-@each $id, $color in $brand-palette {
-  .bg-#{$id} {background-color: nth($color, 1);}
-  .bg-#{$id}-high {background-color: nth($color, 2);}
-  .bg-#{$id}-low {background-color: nth($color, 3);}
-  .bg-#{$id}-gradient {background: radial-gradient(farthest-corner at 0% 0%, nth($color, 2), nth($color, 3));}
-}
