--- src/css/_color.scss
@@ -1,13 +1,0 @@
-@each $id, $color in $ui-palette {
-  .color-#{$id} { color: $color; }
-  .fill-#{$id} {fill: $color;}
-}
-
-@each $id, $color in $brand-palette {
-  .color-#{$id} {color: nth($color, 1);}
-  .color-#{$id}-high {color: nth($color, 2);}
-  .color-#{$id}-low {color: nth($color, 3);}
-  .fill-#{$id} {fill: nth($color, 1);}
-  .fill-#{$id}-high {fill: nth($color, 2);}
-  .fill-#{$id}-low {fill: nth($color, 3);}
-}
