--- src/css/_config.scss
@@ -1,71 +1,0 @@
-//Border
-$border-width: 4px;
-$border-radius: 4px;
-$border-pill: 16px;
-
-//Breakpoints
-$breakpoint-xs: '(min-width: 32rem)';
-$breakpoint-sm: '(min-width: 40rem)';
-$breakpoint-md: '(min-width: 52rem)';
-$breakpoint-lg: '(min-width: 64rem)';
-
-// Loop Counts
-$margin-count: 8; //Generates margin classes 0.5rem increments
-$padding-count: 8; //Generates padding classes in 0.5rem increments
-$layout-count: 25; //Generates height/width classes in 0.5rem increments
-
-//Type Scale
-$h1: 2rem;    // 32px
-$h2: 1.5rem;   // 24px
-$h3: 1.25rem;  // 20px
-$h4: 1rem;      // 16px
-$h5: 0.875rem;  // 14px
-$h6: 0.75rem;   // 12px
-
-$line-height-1:  1;
-$line-height-2:  1.25;
-$line-height-3:  1.5;
-$line-height-4:  2;
-
-// UI Color Definitions
-$white: #FFFFFF;
-$black: #000000;
-$spectre-gray: #171A21;
-$phantom-gray: #43484F;
-$wraith-gray: #9BA6AF;
-$ghost-gray: #E1ECF2;
-$shade-gray: #F5F6F7;
-
-// UI Palette
-$ui-palette: (
-  primary: $black,
-  secondary: $phantom-gray,
-  tertiary: $wraith-gray,
-  placeholder: $ghost-gray,
-  light: $white,
-);
-
-// Brand Color Definitions
-$brand: #9065FF;
-  $brand-high: #7E16FF;
-  $brand-low: #3818E5;
-
-$success: #26D58F;
-  $success-high: #28E398;
-  $success-low: #24C987;
-
-$attention: #EBCD2E;
-  $attention-high: #F7D830;
-  $attention-low: #DEC22B;
-
-$danger: #E91C53;
-  $danger-high: #F51D57;
-  $danger-low: #DB1A4E;
-
-// Brand Palette
-$brand-palette: (
-  brand:($brand, $brand-high, $brand-low),
-  success:($success, $success-high, $success-low),
-  attention:($attention, $attention-high, $attention-low),
-  danger:($danger, $danger-high, $danger-low),
-);
