--- src/css/_hide.scss
@@ -1,26 +1,0 @@
-.hide {
-  position: absolute !important;
-  height: 1px;
-  width: 1px;
-  overflow: hidden;
-  clip: rect(1px, 1px, 1px, 1px);
-}
-
-.invisible {
-  visibility: hidden;
-}
-
-@media #{$breakpoint-sm} {
-  .sm-hide { display: none !important }
-}
-
-@media #{$breakpoint-md} {
-  .md-hide { display: none !important }
-}
-
-@media #{$breakpoint-lg} {
-  .lg-hide { display: none !important }
-}
-
-.display-none { display: none !important }
-
