--- src/css/_hover.scss
@@ -1,13 +1,0 @@
-.clickable {
-  cursor: pointer;
-  user-select: none;
-}
-
-.hover-shadow {
-  transition: box-shadow 0.5s 0.05s;
-  box-shadow: 0 0 0 rgba($black, .2);
-
-  &:hover {
-    box-shadow: 0 0.5rem 2rem rgba($black, .2);
-  }
-}
