--- src/css/_opacity.scss
@@ -1,3 +1,0 @@
-@for $i from 1 through 10 {
-  .o#{$i * 10} { opacity: #{$i * .1}; }
-}
