--- src/css/_position.scss
@@ -1,17 +1,0 @@
-.relative { position: relative; }
-.absolute { position: absolute; }
-.fixed { position: fixed; }
-
-.top-0 { top: 0; }
-.right-0 { right: 0; }
-.bottom-0 { bottom: 0; }
-.left-0 { left: 0; }
-
-.cover {
-  background-size: cover;
-  object-fit: cover;
-}
-
-@for $i from 1 through 10 {
-  .z#{$i} {z-index: #{$i};}
-}
