--- src/css/_reset.scss
@@ -1,37 +1,0 @@
-* {
-  border: 0;
-  box-sizing: border-box;
-  -webkit-font-smoothing: antialiased;
-  font-weight: normal;
-  margin: 0;
-  outline: 0;
-  padding: 0;
-  text-decoration: none;
-  text-rendering: optimizeLegibility;
-  // list-style: none;
-}
-
-html,
-body {
-  min-height: 100vh;
-  min-width: 100%;
-}
-
-a {
-  color: currentColor;
-  text-decoration: none;
-}
-
-textarea {
-  resize: none;
-}
-
-::selection {
-  background: $brand;
-  color: $white;
-}
-
-input,
-textarea {
-  box-shadow: inset 0 1px 1px rgba($black, .1);
-}
