--- src/css/_typography.scss
@@ -1,51 +1,0 @@
-//Fonts
-$body: 'SF UI Text', 'San Francisco', -apple-system, BlinkMacSystemFont, sans-serif;
-
-.font-body {
-  font-family: $body;
-}
-
-.h1 { font-size: $h1; }
-.h2 { font-size: $h2; }
-.h3 { font-size: $h3; }
-.h4 { font-size: $h4; }
-.h5 { font-size: $h5; }
-.h6 { font-size: $h6; }
-
-.font-family-inherit { font-family: inherit; }
-.font-size-inherit { font-size: inherit; }
-.text-decoration-none { text-decoration: none; }
-
-.bold { font-weight: 700; }
-.medium { font-weight: 500; }
-.regular { font-weight: 400; }
-.italic { font-style: italic; }
-.caps { text-transform: uppercase; }
-
-.left-align { text-align: left; }
-.center { text-align: center; }
-.right-align { text-align: right; }
-.justify { text-align: justify; }
-
-.nowrap { white-space: nowrap; }
-.break-word { word-wrap: break-word; }
-
-.line-height-1 { line-height: $line-height-1; }
-.line-height-2 { line-height: $line-height-2; }
-.line-height-3 { line-height: $line-height-3; }
-.line-height-4 { line-height: $line-height-4; }
-
-.list-style-none { list-style: none; }
-.underline { text-decoration: underline; }
-
-.truncate {
-  max-width: 100%;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.list-reset {
-  list-style: none;
-  padding-left: 0;
-}
