--- src/data/posts.js
@@ -1,32 +1,0 @@
-var posts = [
-	{
-		"post": {
-			"id": "1",
-			"user": "Spectrum",
-			"body": {
-				"url": "http://spec.fm/spectrum",
-				"text": "Welcome to Spectrum!"
-			},
-			"chat": [
-				{
-					"message": {
-						"id": "1",
-						"body": "ugh. another design tool.",
-						"user": "Brian Lovin"
-					}
-				},
-				{
-					"message": {
-						"id": "2",
-						"body": "hey man. fuck you.",
-						"user": "Bryn Jackson"
-					}
-				}
-				//messages...
-			]
-		}
-	}
-	//posts...
-];
-
-export default posts;
