+++ src/index.js
@@ -1,17 +1,13 @@
 import React from 'react';
 import { render } from 'react-dom';
-import App from './components/App';
+import App from './App';
 import { BrowserRouter, Match } from 'react-router';
-
-import './css/spector.css';
 
 
 const Root = () => {
 	return(
 		<BrowserRouter>
-			<div>
 				<Match exactly pattern="/" component={App}/>
-			</div>
 		</BrowserRouter>
 	)
 }
